fix: 최근본 페이지 라우트 수정
